// src/lib/types/product.ts
export interface RecommendedProduct {
  title: string;
  link: string;
  image: string;
  lprice: number;
  mallName: string;
  hprice?: string;
  productId?: string;
  productType?: string;
  brand?: string;
  maker?: string;
  category1?: string;
  category2?: string;
  category3?: string;
  category4?: string;
}